Split up the execution code into multiple files
