+++ .github/workflows/build.yml
@@ -60,5 +60,5 @@ jobs:
 
           set -x
           cargo fmt --check
-          cargo clippy
+          cargo clippy -- --deny "warnings"
           cargo test
