+++ flake.nix
@@ -36,7 +36,6 @@
           buildInputs = rix-deps ++ nixjs-rt-deps;
           shellHook = ''
             export RIX_NIXRT_JS_MODULE=nixjs-rt/dist/lib.mjs
-            export RUSTFLAGS=-Dwarnings
           '';
         };
       });
