+++ src/cmd/transpile.rs
@@ -1,5 +1,5 @@
 use crate::cmd::{to_cmd_err, RixSubCommand};
-use crate::eval::nix_v8;
+use crate::eval::execution;
 use clap::{Arg, ArgAction, ArgMatches};
 
 pub fn cmd() -> RixSubCommand {
@@ -26,7 +26,7 @@ pub fn handle_cmd(parsed_args: &ArgMatches) -> Result<(), String> {
         .ok_or("You must provide a single expression to transpile.")?;
     let is_expression = parsed_args.get_one::<bool>("expr").unwrap();
     if *is_expression {
-        let js_source = nix_v8::emit_module(expression)
+        let js_source = execution::emit_module(expression)
             .map_err(|_err| "Failed to transpile the expression.".to_owned())?;
         print!("{js_source}");
     } else {
