+++ src/eval/mod.rs
@@ -1,2 +1,5 @@
-pub mod nix_v8;
+pub mod emit_js;
+pub mod execution;
+pub mod helpers;
+pub mod tests;
 pub mod types;
